feat: support small heights (#153)
Proof system supports arbitrary log trace heights now, even those less
than `l_skip`.

Closes INT-5092

Note for reviewers:
- If you are working on AIRs, only look at verifier changes.
- If working on cuda backend, only look at prover changes.

# Recursion

## Buses Changed
* (New) HyperdimBus
* (Changed) AirHeightsBus -> LiftedHeightsBus

```rust
// Proof shape <-> SymbolicExpressionAir
pub struct HyperdimBusMessage<T> {
    pub sort_idx: T,
    pub n_abs: T,
    pub n_sign_bit: T,
}

// Proof shape <-> StackingReduction
pub struct LiftedHeightsBusMessage<T> {
    pub sort_idx: T,
    /// Sender must constraint this equals `2^log_lifted_height`.
    pub lifted_height: T,
    /// Sender must constrain this equals `max(log_height, l_skip)`.
    pub log_lifted_height: T,
}
```

---------

+++ crates/stark-backend-v2/src/v1_shims.rs
@@ -59,7 +59,7 @@ impl StarkProvingKeyV2 {
             );
             preprocessed_vdata = Some(VerifierSinglePreprocessedData {
                 commit,
-                hypercube_dim: log2_strict_usize(trace.height()) - params.l_skip,
+                hypercube_dim: log2_strict_usize(trace.height()) as isize - params.l_skip as isize,
                 stacking_width: data.matrix.width(),
             });
             Arc::new(data)
